Initial commit of the new LGP project. This version of the LGP Project is a total rebuild that may replace the initial version of the LGP library.
I am doing this for a few reasons.  First, it is looking chaotic
enough that I must be going about it wrong.  Maybe not, we will
see.  Second, I am wanting to take a fresh look at the project
and reorganize things.

So far, I have a skeleton directory set up, and a new Makefile. The
first thing I have done from the start is to try to organize the
Makefile in a manner that will be more readable overall.