I have imported some pieces of the old lgp to get started.
I imported the array, stack, and random numbers portions
of the old libgamepieces to get me started.  these are
things that I know I will reuse and am not as likely to
make massive changes too either way.  I would be redoing
a lot of work if I just did them over.

The main changes that I had to make to get them to compile
with the new environment and the new style make file was to
change the include lines from #include <lgp_module> to
say #include <lgp/lgp_module> instead.