Started fleshing out lgp::Object and lgp::GameSystem.
I began to work a little bit on the lgp::Object interface
as well as created the header file for lgp::GameSystem. I
also created a text file, doc/GameSystemNotes.txt, to have
a space to brainstorm a bit on what a person might require for
a minimal game system.  I will also try to get some friends
to brainstorm on the subject as well.