Refactoring and constructors.
Renamed lgp::Object::id() and lgp::Object::type() to get_id() and get_type()
so that they would be more in line with what is considered normal naming
for setters and getters.

I also created a new constructor for lgp::Object to let us set the basic
data when the object is created as well as a little bit of tweaking in
lgp::ObjectData.